+++ src/client/profile-provider.ts
@@ -73,7 +73,7 @@ export class BoundProfileProvider {
   constructor(
     private readonly profileAst: ProfileDocumentNode,
     private readonly mapAst: MapDocumentNode,
-    private readonly providerName: string,
+    private readonly provider: ProviderJson,
     readonly configuration: {
       services: IServiceSelector;
       profileProviderSettings?: NormalizedProfileProviderSettings;
@@ -87,7 +87,7 @@ export class BoundProfileProvider {
     this.fetchInstance = new CrossFetch();
     this.fetchInstance.metadata = {
       profile: profileAstId(profileAst),
-      provider: providerName,
+      provider: provider.name,
     };
     this.fetchInstance.events = events;
   }
@@ -122,12 +122,13 @@ export class BoundProfileProvider {
   >(
     usecase: string,
     input?: TInput,
-    parameters?: Record<string, string>
+    parameters?: Record<string, string>,
+    securityValues?: SecurityValues[]
   ): Promise<Result<TResult, ProfileParameterError | MapInterpreterError>> {
     this.fetchInstance.metadata = {
       profile: profileAstId(this.profileAst),
       usecase,
-      provider: this.providerName,
+      provider: this.provider.name,
     };
     // compose and validate the input
     const composedInput = this.composeInput(usecase, input);
@@ -142,13 +143,21 @@ export class BoundProfileProvider {
     }
     forceCast<TInput>(composedInput);
 
+    const security = securityValues
+      ? resolveSecurityConfiguration(
+          this.provider.securitySchemes ?? [],
+          securityValues,
+          this.provider.name
+        )
+      : this.configuration.security;
+
     // create and perform interpreter instance
     const interpreter = new MapInterpreter<TInput>(
       {
         input: composedInput,
         usecase,
         services: this.configuration.services,
-        security: this.configuration.security,
+        security,
         parameters: this.mergeParameters(
           parameters,
           this.configuration.parameters
@@ -344,7 +353,7 @@ export class ProfileProvider {
     return new BoundProfileProvider(
       profileAst,
       mapAst,
-      providerInfo.name,
+      providerInfo,
       {
         services: new ServiceSelector(
           providerInfo.services,
@@ -672,24 +681,19 @@ export class ProfileProvider {
   /**
    * Resolves auth variables by applying the provided overlay over the base variables.
    *
-   * The base variables either come from super.json or from `this.provider` if it is an instance of `ProviderConfiguration`
+   * The base variables come from super.json
    */
   private resolveSecurityValues(
     providerName: string,
     overlay?: SecurityValues[]
   ): SecurityValues[] {
-    let base: SecurityValues[];
-    if (this.provider instanceof ProviderConfiguration) {
-      base = this.provider.security;
-    } else {
-      base = this.superJson.normalized.providers[providerName]?.security;
-    }
+    const base: SecurityValues[] =
+      this.superJson.normalized.providers[providerName]?.security ?? [];
 
-    let resolved = base;
     if (overlay !== undefined) {
-      resolved = mergeSecurity(base, overlay);
+      return mergeSecurity(base, overlay);
     }
 
-    return resolved;
+    return base;
   }
 }